examples/usage_from_python_script/pyproject.toml
[tool.poetry]
name = "usage-from-python-script"
version = "0.1.0"
description = "This example demonstrates how to use the SDK to run a simple app from within a Python script."
authors = []
license = "MIT"
readme = "README.md"
repository = "https://github.com/otherwillhq/scoutos-python/examples/usage_from_python_script"


[tool.poetry.dependencies]
python = "^3.11"
scoutos = { path = "../..", develop = true }
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
